refactor: Update ProductController and Product model

+++ app/Http/Controllers/ProductController.php
@@ -6,20 +6,16 @@ use App\Models\Product;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use App\Http\Requests\StoreProductRequest;
+use Illuminate\Support\Facades\Gate;
 
 class ProductController extends Controller
 {
     /**
      * Instantiate a new controller instance.
-     *
-     * @return void
      */
     public function __construct()
     {
-        $this->middleware('auth')->except(['index', 'show']);
-
-        // This controller methods will be mapped to their corresponding policy method.
-        $this->authorizeResource(Product::class, 'product');
+        $this->middleware('auth', ['except' => ['index', 'show']]);
     }
 
     /**
@@ -29,11 +25,9 @@ class ProductController extends Controller
      */
     public function index()
     {
-        $products = Product::where('status', '=', 'public')->get();
+        $products = Product::where('status', 'public')->get();
 
-        return view('products.index', [
-            'products' => $products
-        ]);
+        return view('products.index', compact(['products']));
     }
 
     /**
@@ -43,6 +37,8 @@ class ProductController extends Controller
      */
     public function create()
     {
+        Gate::authorize('create', Product::class);
+
         return view('products.create');
     }
 
@@ -54,6 +50,8 @@ class ProductController extends Controller
      */
     public function store(StoreProductRequest $request)
     {
+        Gate::authorize('create', Product::class);
+
         $data = $request->validated();
 
         $product = new Product;
@@ -75,9 +73,9 @@ class ProductController extends Controller
      */
     public function show(Product $product)
     {
-        return view('products.show', [
-            'product' => $product
-        ]);
+        Gate::authorize('view', $product);
+
+        return view('products.show', compact(['product']));
     }
 
     /**
@@ -87,9 +85,9 @@ class ProductController extends Controller
      */
     public function edit(Product $product)
     {
-        return view('products.edit', [
-            'product' => $product
-        ]);
+        Gate::authorize('update', $product);
+
+        return view('products.edit', compact(['product']));
     }
 
     /**
@@ -100,16 +98,15 @@ class ProductController extends Controller
      */
     public function update(StoreProductRequest $request, Product $product)
     {
+        Gate::authorize('update', $product);
+
         $data = $request->validated();
 
         $product->update($data);
-        $product->save();
 
         $request->session()->flash('success', 'Produto atualizado!');
 
-        return redirect()->route('products.show', [
-            'product' => $product->slug
-        ]);
+        return redirect()->route('products.show', ['product' => $product->slug]);
     }
 
     /**
@@ -119,6 +116,8 @@ class ProductController extends Controller
      */
     public function destroy(Request $request, Product $product)
     {
+        Gate::authorize('delete', $product);
+
         $product->delete();
 
         $request->session()->flash('success', 'Produto removido com sucesso!');
